+++ service/controller/resource/deletecrs/create.go
@@ -1,4 +1,4 @@
-package deleteconfigured
+package deletecrs
 
 import (
 	"context"
